requirements.txt
flask-restplus==0.9.2
Werkzeug==0.16.1
gunicorn==20.0.4
flask-mongoengine==1.0.0
mongoengine==0.23.0
pymongo==3.11.3
prometheus-client==0.8.0
prometheus-flask-exporter==0.15.4
pika==1.2.0
